pyproject.toml
[tool.poetry]
name = "freidok"
version = "0.1.0"
description = ""
authors = ["Alex <alexander.haupt@physiologie.uni-freiburg.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.10.*"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
pydantic = "^1.10.4"
lxml = "^4.9.2"

[tool.poetry.scripts]
freidok = 'freidok.cli:run'

[tool.poetry.group.dev.dependencies]
datamodel-code-generator = {extras = ["http"], version = "^0.15.0"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
